requirements.txt
asgiref==3.8.1
Django>=3.2,<4.0
django-cors-headers==4.3.1
django-rest-authtoken==2.1.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
tzdata==2024.1
requests
Pillow
djangorestframework-simplejwt